Separate checks into a module
